Fix: Vulnerability in Comparison of different type sizes #1001

+++ src/PubSubClient.cpp
@@ -457,7 +457,7 @@ boolean PubSubClient::publish(const char* topic, const uint8_t* payload, unsigne
         length = writeString(topic,this->buffer,length);
 
         // Add payload
-        uint16_t i;
+        unsigned int i;
         for (i=0;i<plength;i++) {
             this->buffer[length++] = payload[i];
         }
